+++ src/modules/newCoins/content/newCoins.tsx
@@ -1,0 +1,6 @@
+import { NewCoinsStyle } from "../style/newCoins_styles";
+import React from "react";
+
+export default function NewCoins() {
+  return <NewCoinsStyle>NewCoins</NewCoinsStyle>;
+}
