+++ src/modules/newCoins/style/newCoins_styles.ts
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export const NewCoinsStyle = styled.div(({}) => ({}));
